Flot is now integrated quite good
Added a rendering class to seperate the data that needs to be generated for
highcharts and for flot. Currently only a FLOT renderer is written. I will
do the same on the master branch.
ww_gpx_info.php is changed so the flot-renderer is utilized. Graphs drawn
by flot already look similar to highchart graphs. Layout of legend needs to
be done. I also have to add the interactive-features and formatters.