Change enqueue of scripts
Filter the posts first to find out if our shortcode is present.
Then enqueue the libraries that we need.
Further we load the minimized version of each java-script by default if
debugging is not enabled in the class during init process
We also check if the highcharts-api is downloaded and installed before
trying to include it.

Signed-off-by: Walter Werther <walter@wwerther.de>